+++ payoneer-checkout/modules/inpsyde/payoneer-embedded-payment/assets/payoneer-checkout.asset.php
@@ -2,4 +2,4 @@
 
 namespace Syde\Vendor;
 
-return array('dependencies' => array('jquery', 'react', 'wc-blocks-data-store', 'wp-api-fetch', 'wp-data', 'wp-hooks'), 'version' => 'e1e588052347528f5d49');
+return array('dependencies' => array('jquery', 'react', 'wc-blocks-data-store', 'wp-api-fetch', 'wp-data', 'wp-hooks'), 'version' => 'f53ba10c05c650d8bf28');