+++ payoneer-checkout/modules/inpsyde/payoneer-embedded-payment/src/PaymentProcessor/EmbeddedPaymentProcessor.php
@@ -1,42 +1,34 @@
 <?php
 
-declare (strict_types=1);
 namespace Syde\Vendor\Inpsyde\PayoneerForWoocommerce\EmbeddedPayment\PaymentProcessor;
 
 use Exception;
-use Syde\Vendor\Inpsyde\PayoneerForWoocommerce\Checkout\Authentication\TokenGeneratorInterface;
-use Syde\Vendor\Inpsyde\PayoneerForWoocommerce\Checkout\MisconfigurationDetector\MisconfigurationDetectorInterface;
-use Syde\Vendor\Inpsyde\PayoneerForWoocommerce\Checkout\PaymentProcessor\AbstractPaymentProcessor;
-use Syde\Vendor\Inpsyde\PayoneerForWoocommerce\ListSession\ListSession\ListSessionProvider;
-use Syde\Vendor\Inpsyde\PayoneerForWoocommerce\ListSession\ListSession\PaymentContext;
-use Syde\Vendor\Inpsyde\PayoneerForWoocommerce\Api\Gateway\CommandFactory\WcOrderBasedUpdateCommandFactoryInterface;
 use Syde\Vendor\Inpsyde\PaymentGateway\PaymentGateway;
+use Syde\Vendor\Inpsyde\PaymentGateway\PaymentProcessorInterface;
+use Syde\Vendor\Inpsyde\PayoneerForWoocommerce\PaymentMethods\PaymentProcessor\PayoneerCommonPaymentProcessor;
 use Syde\Vendor\Inpsyde\PayoneerSdk\Api\Command\Exception\CommandExceptionInterface;
 use Syde\Vendor\Inpsyde\PayoneerSdk\Api\Command\Exception\InteractionExceptionInterface;
 use WC_Order;
 /**
- * @psalm-import-type PaymentResult from AbstractPaymentProcessor
+ * @psalm-import-type PaymentResult from PayoneerCommonPaymentProcessor
  */
-class EmbeddedPaymentProcessor extends AbstractPaymentProcessor
+class EmbeddedPaymentProcessor implements PaymentProcessorInterface
 {
-    /**
-     * @var string
-     */
-    protected string $hostedModeOverrideFlag;
-    protected bool $isRestRequest;
-    public function __construct(WcOrderBasedUpdateCommandFactoryInterface $updateCommandFactory, ListSessionProvider $sessionProvider, TokenGeneratorInterface $tokenGenerator, string $tokenKey, string $transactionIdFieldName, string $hostedModeOverrideFlag, MisconfigurationDetectorInterface $misconfigurationDetector, string $checkoutSessionHashKey, bool $isRestRequest)
+    private PayoneerCommonPaymentProcessor $commonProcessor;
+    private string $hostedModeOverrideFlag;
+    private bool $isRestRequest;
+    public function __construct(PayoneerCommonPaymentProcessor $commonProcessor, string $hostedModeOverrideFlag, bool $isRestRequest)
     {
-        parent::__construct($misconfigurationDetector, $sessionProvider, $updateCommandFactory, $tokenGenerator, $tokenKey, $transactionIdFieldName, $checkoutSessionHashKey);
+        $this->commonProcessor = $commonProcessor;
         $this->hostedModeOverrideFlag = $hostedModeOverrideFlag;
         $this->isRestRequest = $isRestRequest;
     }
     public function processPayment(WC_Order $order, PaymentGateway $gateway): array
     {
-        $this->sessionProvider->provide(new PaymentContext($order));
         try {
-            $result = parent::processPayment($order, $gateway);
+            $result = $this->commonProcessor->processPayment($order, $gateway);
         } catch (InteractionExceptionInterface $exception) {
-            return $this->handleInteractionException($order, $exception);
+            return $this->commonProcessor->handleInteractionException($order, $exception);
         } catch (CommandExceptionInterface $exception) {
             $exceptionWrapper = new Exception(
                 /* translators: An unexpected error during the final List UPDATE before the CHARGE */
@@ -44,7 +36,7 @@ class EmbeddedPaymentProcessor extends AbstractPaymentProcessor
                 $exception->getCode(),
                 $exception
             );
-            return $this->handleFailedPaymentProcessing($order, $exceptionWrapper);
+            return $this->commonProcessor->handleFailedPaymentProcessing($order, $exceptionWrapper);
         }
         /**
          * We always signal success: The actual payment is supposed to be handled by the JS WebSDK
@@ -61,7 +53,7 @@ class EmbeddedPaymentProcessor extends AbstractPaymentProcessor
         }
         /* translators: Order note added when processing an order in embedded flow */
         $note = __('Backend processing finished, frontend processing is about to start.', 'payoneer-checkout');
-        $this->putOrderOnHold($order, $note . \PHP_EOL);
+        $this->commonProcessor->putOrderOnHold($order, $note . \PHP_EOL);
         return $result;
     }
 }